266: Testing Card Failures

+++ eCart.SPA/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -67,10 +67,26 @@ export class CheckoutPaymentComponent implements OnInit{
       this.checkoutService.createOrder(orderToCreate).subscribe({
         next: order => {
           this.toastr.success('Order Created Successfully');
-          this.basketService.deleteLocalBasket();    
-          const navigationExtras: NavigationExtras = {state: order};
-          this.router.navigate(['checkout/success'], navigationExtras);
-          console.log(order);      
+
+          this.stripe?.confirmCardPayment(basket.clientSecret!, {
+            payment_method: {
+              card: this.cardNumber!,
+              billing_details: {
+                name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')?.value
+              }
+            }
+          }).then(result => {
+            console.log(result);
+            if(result.paymentIntent) {
+              this.basketService.deleteLocalBasket();    
+              const navigationExtras: NavigationExtras = {state: order};
+              this.router.navigate(['checkout/success'], navigationExtras);    
+            }
+            else {
+              this.toastr.error(result.error.message);
+            }
+          })
+
         }
       });
     }
